requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
discord==2.3.2
discord.py==2.6.4
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2
yarl==1.22.0
